+++ config/urls.py
@@ -6,7 +6,6 @@ from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
 from v1.contributors.urls import router as contributor_router
-from v1.meta.urls import router as meta_router
 from v1.openings.urls import router as opening_router
 from v1.tasks.urls import router as task_router
 from v1.teams.urls import router as team_router
@@ -26,7 +25,6 @@ urlpatterns = [
 router = DefaultRouter(trailing_slash=False)
 
 router.registry.extend(contributor_router.registry)
-router.registry.extend(meta_router.registry)
 router.registry.extend(opening_router.registry)
 router.registry.extend(task_router.registry)
 router.registry.extend(team_router.registry)
