--- v1/meta/admin.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.contrib import admin
-
-from .models import Responsibility, Skill
-
-admin.site.register(Responsibility)
-admin.site.register(Skill)
