--- v1/meta/apps.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-from django.apps import AppConfig
-
-
-class MetaConfig(AppConfig):
-    name = 'v1.meta'
