--- v1/meta/factories/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from .responsibility import ResponsibilityFactory
-from .skill import SkillFactory
-
-__all__ = [
-    'ResponsibilityFactory',
-    'SkillFactory'
-]
