--- v1/meta/models/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from .responsibility import Responsibility
-from .skill import Skill
-
-__all__ = [
-    'Responsibility',
-    'Skill'
-]
