--- v1/meta/serializers/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from .responsibility import ResponsibilitySerializer
-from .skill import SkillSerializer
-
-__all__ = [
-    'ResponsibilitySerializer',
-    'SkillSerializer'
-]
