--- v1/meta/views/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from .responsibility import ResponsibilityViewSet
-from .skill import SkillViewSet
-
-__all__ = [
-    'ResponsibilityViewSet',
-    'SkillViewSet'
-]
