+++ v1/openings/admin.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 from django.contrib import admin
 
-from .models import Opening
+from .models import Opening, Responsibility, Skill
 
 admin.site.register(Opening)
+admin.site.register(Responsibility)
+admin.site.register(Skill)
