+++ v1/openings/factories/__init__.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
 from .opening import OpeningFactory
+from .responsibility import ResponsibilityFactory
+from .skill import SkillFactory
 
 __all__ = [
-    'OpeningFactory'
+    OpeningFactory,
+    ResponsibilityFactory,
+    SkillFactory
 ]
