+++ v1/openings/factories/opening.py
@@ -3,9 +3,10 @@ import factory
 from factory.django import DjangoModelFactory
 
 from v1.teams.factories import TeamFactory
+from .responsibility import ResponsibilityFactory
+from .skill import SkillFactory
 from ..models import Opening
 from ...contributors.factories import ContributorFactory
-from ...meta.factories import ResponsibilityFactory, SkillFactory
 
 
 class OpeningFactory(DjangoModelFactory):
