+++ v1/openings/models/__init__.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
 from .opening import Opening
-
+from .responsibility import Responsibility
+from .skill import Skill
 
 __all__ = [
-    'Opening'
+    Opening,
+    Responsibility,
+    Skill
 ]
