+++ v1/openings/models/opening.py
@@ -6,13 +6,15 @@ from thenewboston.models.created_modified import CreatedModified
 
 from v1.contributors.models import Contributor
 from v1.teams.models import Team
+from .responsibility import Responsibility
+from .skill import Skill
 
 
 class Opening(CreatedModified):
     uuid = models.UUIDField(default=uuid.uuid4, editable=False, primary_key=True)
     reports_to = models.ManyToManyField(Contributor)
-    responsibilities = models.ManyToManyField('meta.Responsibility')
-    skills = models.ManyToManyField('meta.Skill')
+    responsibilities = models.ManyToManyField(Responsibility)
+    skills = models.ManyToManyField(Skill)
     team = models.ForeignKey(Team, on_delete=models.CASCADE)
 
     active = models.BooleanField(default=True, db_index=True)
