+++ v1/openings/models/skill.py
