+++ v1/openings/serializers/__init__.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
 from .opening import OpeningSerializer
-
+from .responsibility import ResponsibilitySerializer
+from .skill import SkillSerializer
 
 __all__ = [
-    'OpeningSerializer',
+    OpeningSerializer,
+    ResponsibilitySerializer,
+    SkillSerializer,
 ]
