+++ v1/openings/serializers/responsibility.py
