+++ v1/openings/serializers/skill.py
