+++ v1/openings/views/__init__.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
 from .opening import OpeningViewSet
-
+from .responsibility import ResponsibilityViewSet
+from .skill import SkillViewSet
 
 __all__ = [
-    'OpeningViewSet'
+    OpeningViewSet,
+    ResponsibilityViewSet,
+    SkillViewSet
 ]
