+++ v1/openings/views/responsibility.py
