+++ v1/openings/views/skill.py
